Move tabs and ol_confirm_dialog to logical place
1) ol_confirm_dialog relates to dialogs so should be inside dialog.js as a new
function confirmDialog
The behaviour for ol_confirm_dialog is modified to only initialise the dialog
inside the click handler.

2) the code inside automatic initialises some tabs so should be
inside tabs

* tabs:not(.ui-tabs) selector is moved from automatic.js to call
to initTabs
* autohash logic is moved from automatic.js moved to tabs.js
($.support.opacity check is unnecessary as its always supported on our
browser matrix)
