+++ openlibrary/plugins/openlibrary/js/automatic.js
@@ -2,22 +2,8 @@
  * Setup actions on document.ready for standard classNames.
  */
 export default function($) {
-    var options;
     // Flash messages are hidden by default so that CSS is not on the critical path.
     $('.flash-messages').show();
-
-    // tabs
-    options = {};
-    if ($.support.opacity){
-        options.fx = {opacity: 'toggle'};
-    }
-
-    if ($('.tabs:not(.ui-tabs)').tabs) {
-        $('.tabs:not(.ui-tabs)').tabs(options)
-        $('.tabs.autohash').bind('tabsselect', function(event, ui) {
-            document.location.hash = ui.panel.id;
-        });
-    }
 
     // validate forms
     $('form.validate').ol_validate();
