+++ openlibrary/plugins/openlibrary/js/index.js
@@ -27,13 +27,13 @@ import * as Browser from './Browser';
 import { commify } from './python';
 import { Subject, urlencode, slice } from './subjects';
 import Template from './template.js';
-// Add $.fn.focusNextInputField, $.fn.ol_confirm_dialog
+// Add $.fn.focusNextInputField
 import { closePopup, truncate, cond } from './utils';
 import initValidate from './validate';
 import '../../../../static/css/js-all.less';
 // polyfill Promise support for IE11
 import Promise from 'promise-polyfill';
-import initDialogs from './dialog';
+import { confirmDialog, initDialogs } from './dialog';
 import initTabs from './tabs.js';
 
 // Eventually we will export all these to a single global ol, but in the mean time
@@ -74,7 +74,9 @@ jQuery(function () {
     // Live NodeList is cast to static array to avoid infinite loops
     const $carouselElements = $('.carousel--progressively-enhanced');
     initDialogs();
-    initTabs($('#tabsAddbook,#tabsAddauthor'));
+    // expose ol_confirm_dialog method
+    $.fn.ol_confirm_dialog = confirmDialog;
+    initTabs($('#tabsAddbook,#tabsAddauthor,.tabs:not(.ui-tabs)'));
     initValidate($);
     autocompleteInit($);
     addNewFieldInit($);
