+++ openlibrary/plugins/openlibrary/js/utils.js
@@ -10,27 +10,6 @@ $.fn.focusNextInputField = function() {
     });
 };
 
-// Confirm dialog with OL styles.
-$.fn.ol_confirm_dialog = function(callback, options) {
-    var _this = this;
-    var defaults = {
-        autoOpen: false,
-        width: 400,
-        modal: true,
-        resizable: false,
-        buttons: {
-            'Yes, I\'m sure': function() {
-                callback.apply(_this);
-            },
-            'No, cancel': function() {
-                $(_this).dialog('close');
-            }
-        }
-    };
-    options = $.extend(defaults, options);
-    this.dialog(options);
-}
-
 // closes active popup
 // used in templates/covers/saved.html
 export function closePopup() {
