+++ openlibrary/templates/lists/lists.html
@@ -27,21 +27,20 @@ $if "type" in doc and doc.type.key == "/type/user":
                     .find(".listDelete a")
                     .click(function() {
                         \$('#delete-dialog')
+                            .ol_confirm_dialog(function(){
+                                var list_id = "#" + \$(this).data("list-id");
+                                var key = \$(list_id + " a:first").attr("href");
+                                var dialog = this;
+
+                                \$.post(key + "/delete.json", function() {
+                                    \$(list_id).remove();
+                                    \$(dialog).dialog("close");
+                                });
+                            })
                             .data("list-id", \$(this).closest("li").attr("id"))
                             .dialog('open');
                         return false;
                     });
-
-                \$("#delete-dialog").ol_confirm_dialog(function(){
-                    var list_id = "#" + \$(this).data("list-id");
-                    var key = \$(list_id + " a:first").attr("href");
-                    var dialog = this;
-
-                    \$.post(key + "/delete.json", function() {
-                        \$(list_id).remove();
-                        \$(dialog).dialog("close");
-                    });
-                });
             });
         </script>
         <div id="delete-dialog" class="hidden" title="Delete list">Are you sure you want to delete this list?</div>
@@ -56,22 +55,21 @@ $else:
             .find(".listDelete a")
             .click(function() {
                 \$('#remove-dialog')
+                    .ol_confirm_dialog(function() {
+                        var seed = $:json_encode(seed);
+                        var list_id = "#" + \$(this).data("list-id");
+                        var key = \$(list_id + " a:first").attr("href");
+
+                        var dialog = this;
+
+                        remove_seed(key, seed, function() {
+                            \$(list_id).remove();
+                            \$(dialog).dialog("×");
+                        });
+                    })
                     .data("list-id", \$(this).closest("li").attr("id"))
                     .dialog('open');
                 return false;
-            });
-
-            \$("#remove-dialog").ol_confirm_dialog(function() {
-                var seed = $:json_encode(seed);
-                var list_id = "#" + \$(this).data("list-id");
-                var key = \$(list_id + " a:first").attr("href");
-
-                var dialog = this;
-
-                remove_seed(key, seed, function() {
-                    \$(list_id).remove();
-                    \$(dialog).dialog("×");
-                });
             });
         });
